+++ app/src/main/java/com/example/myapplication/dto/TransactionDto.kt
@@ -1,11 +1,12 @@
 package com.example.myapplication.dto
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class TransactionDto(
     @SerializedName("id") val id: Long,
     @SerializedName("is_income") val isIncome: Boolean,
-    @SerializedName("amount") val amount: Double,
+    @SerializedName("amount") val amount: BigDecimal,
     @SerializedName("category_id") val categoryId: Long,
     @SerializedName("description") val description: String?,
     @SerializedName("date") val date: String,
