+++ app/src/main/java/com/example/myapplication/schemas/BalanceCardUpdateSchema.kt
@@ -1,6 +1,8 @@
 package com.example.myapplication.schemas
 
+import java.math.BigDecimal
+
 data class BalanceCardUpdateSchema (
     val card_id: Long,
-    val new_balance: Double
+    val new_balance: BigDecimal
 )
